requirements-prod.txt
# TitleCraft AI - Production Dependencies
# Minimal production setup: pip install -r requirements-prod.txt

# Core API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
pydantic>=2.4.0

# Async operations
aiohttp>=3.9.0
httpx>=0.25.0

# Essential data processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Infrastructure
redis>=5.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# LLM providers (core)
openai>=1.3.0
anthropic>=0.7.0

# Production monitoring
prometheus-client>=0.19.0
structlog>=23.2.0
psutil>=5.9.0

# Security
cryptography>=41.0.0

# Optional performance enhancements
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0